feat: add help for getting started steps
