+++ src/App.vue
@@ -55,7 +55,7 @@ h5,
 h6 {
   font-weight: 500;
 }
-@media (max-width: 768px) {
+@media (max-width: 860px) {
   h1 {
     font-size: 1.5rem;
   }
