+++ src/components/CodeBlock.vue
@@ -168,7 +168,7 @@ div[class~='language-yaml']::before {
   content: 'yaml';
 }
 /* media queries */
-@media (max-width: 768px) {
+@media (max-width: 860px) {
   .code-block-wrapper {
     height: 100%;
   }
