--- src/components/IconDownload.vue
@@ -1,18 +1,0 @@
-<template>
-  <svg
-    xmlns="http://www.w3.org/2000/svg"
-    width="1.4em"
-    height="1.4em"
-    viewBox="0 0 24 24"
-    fill="none"
-    stroke="currentColor"
-    stroke-width="2"
-    stroke-linecap="round"
-    stroke-linejoin="round"
-    class="icons feather feather-download"
-  >
-    <path d="M21 15v4a2 2 0 0 1-2 2H5a2 2 0 0 1-2-2v-4"></path>
-    <polyline points="7 10 12 15 17 10"></polyline>
-    <line x1="12" y1="15" x2="12" y2="3"></line>
-  </svg>
-</template>
