+++ src/components/NavTwitter.vue
@@ -1,0 +1,20 @@
+<template>
+  <a
+    class="external-links"
+    href="https://twitter.com/pytorch_ignite"
+    target="_blank"
+    rel="noopener noreferrer"
+  >
+    <svg width="1.4em" height="1.4em" viewBox="0 0 512 512" class="icons">
+      <path
+        d="M459.37 151.716c.325 4.548.325 9.097.325 13.645c0 138.72-105.583 298.558-298.558 298.558c-59.452 0-114.68-17.219-161.137-47.106c8.447.974 16.568 1.299 25.34 1.299c49.055 0 94.213-16.568 130.274-44.832c-46.132-.975-84.792-31.188-98.112-72.772c6.498.974 12.995 1.624 19.818 1.624c9.421 0 18.843-1.3 27.614-3.573c-48.081-9.747-84.143-51.98-84.143-102.985v-1.299c13.969 7.797 30.214 12.67 47.431 13.319c-28.264-18.843-46.781-51.005-46.781-87.391c0-19.492 5.197-37.36 14.294-52.954c51.655 63.675 129.3 105.258 216.365 109.807c-1.624-7.797-2.599-15.918-2.599-24.04c0-57.828 46.782-104.934 104.934-104.934c30.213 0 57.502 12.67 76.67 33.137c23.715-4.548 46.456-13.32 66.599-25.34c-7.798 24.366-24.366 44.833-46.132 57.827c21.117-2.273 41.584-8.122 60.426-16.243c-14.292 20.791-32.161 39.308-52.628 54.253z"
+        fill="currentColor"
+      ></path>
+    </svg>
+    <span class="icon-text">Twitter</span>
+  </a>
+</template>
+
+<style scoped>
+@import url('./css/nav-right.css');
+</style>
