+++ src/components/css/nav-right.css
@@ -1,0 +1,22 @@
+.external-links {
+  display: inline-flex;
+  align-content: center;
+  justify-content: center;
+  margin: 0 0.5rem;
+  border-bottom: 2px solid transparent;
+}
+.external-links:hover {
+  border-bottom: 2px solid var(--c-brand-red);
+}
+.icon-text {
+  margin-left: 0.25rem;
+}
+/* media queries */
+@media (max-width: 860px) {
+  .external-links {
+    margin: 0 0.25rem;
+  }
+  .icon-text {
+    display: none;
+  }
+}
